Move audio-output-related classes to separate .h and .cpp files
- Move PortAudioSystem, AudioStream and AudioOutput classes to audio_output.h and audio_output.cpp
- Update CMakeLists.txt
